detect/http-uri: move tests into tests/
